modelbench/requirements.txt
datasets>=4.8,<5.0
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
mtcnn>=1.0,<2.0
numpy>=1.26,<3.0
pillow>=10.4,<12.0
pytest>=8.0,<9.0
python-multipart>=0.0.9,<1.0
tensorflow>=2.16,<2.20
uvicorn>=0.30,<1.0
